requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
google-genai==0.3.0
Pillow==10.2.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx~=0.26.0